Reworked the database so its more uniform added some buttons started removing users
Had to rewite the meeting setting code since it was written baddly. So now each column is a new person. Add some buttons so the user can set as a recip or init. This is for later so the user can view meeting or create them. Started removing users but found my database to be incorrect so i had to rearrage first. It removes users but all the users atm so adding the data as columns was just easier then reading the database then readding. Its also quicker if the database was large.